+++ packages/ast-utilities/CHANGELOG.md
@@ -7,6 +7,10 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## 0.0.2 - 2019-09-18
+
+- Added markdown utils starting with `addReleaseToChangelog`
+
 ### Added
 
 - `@shopify/ast-utilities` package
