+++ packages/ast-utilities/README.md
@@ -13,38 +13,25 @@ $ yarn add @shopify/ast-utilities
 
 ## Usage
 
-1. [Utilities](#utilities)
-1. [Transforms](#transforms)
+This library provides utilities for transforming and working with ASTs across a set of languages in a common way. We currently support markdown, [using remark](https://remark.js.org), and JavaScript, [using babel](https://babeljs.io/).
 
-### Utilities
+### `transform(code, ...transforms)`
 
-#### `transform(code, ...transforms)`
+Each language provides a `transform` function with the same signature. The first arugment is the initial code, and is followed by any number of transform functions to apply. The resulting string is returned.
 
-This utility applies any number of transforms onto a given block of code and returns the resulting string. See [Transforms](#transforms) examples below.
+1. [JavaScript](#javascript)
+1. [Markdown](#markdown)
 
-#### `astFrom(code)`
+## Javascript
 
-This utility provides a simple way to parse a string into an AST statement.
+`@shopify/ast-transforms/javascript`
 
-```tsx
-import {astFrom} from '@shopify/ast-utilities';
-
-const module = '@shopify/ast-utilities';
-const util = 'compose';
-
-const ast = astFrom`
-  const {${util}} = require("${module}");
-`;
-```
-
-## Transforms
-
-#### `addComponentProps(props, componentName)`
+### `addComponentProps(props, componentName)`
 
 Takes the first argument array of props and adds them to any component name that match the given second argument string.
 
 ```tsx
-import {transform, addComponentProps} from '@shopify/ast-transforms';
+import {transform, addComponentProps} from '@shopify/ast-transforms/javascript';
 
 const initial = `<Foo />`;
 
@@ -59,11 +46,16 @@ const result = await transform(
 console.log(result); // <Foo someProp={someValue} />
 ```
 
-#### `addImportSpecifier(importSource, newSpecifier)`
+### `addImportSpecifier(importSource, newSpecifier)`
 
 Adds an import specifier, or collection of specifiers, to an existing import statement, or create a new import statement with the given specifier if it does not already exist.
 
 ```tsx
+import {
+  transform,
+  addImportSpecifier,
+} from '@shopify/ast-transforms/javascript';
+
 const initial = `import {foo} from './bar'`;
 
 const result = await transform(initial, addImportSpecifier('./bar', 'baz'));
@@ -71,11 +63,16 @@ const result = await transform(initial, addImportSpecifier('./bar', 'baz'));
 console.log(result); // import {foo, baz} from './bar',
 ```
 
-#### `addImportStatement(statements)`
+### `addImportStatement(statements)`
 
 Adds an import statement, or collection of import statements, to a file.
 
 ```tsx
+import {
+  transform,
+  addImportStatement,
+} from '@shopify/ast-transforms/javascript';
+
 const initial = `import {foo} from './bar'`;
 
 const result = await transform(
@@ -88,11 +85,13 @@ const result = await transform(
 console.log(result);
 ```
 
-#### `replaceStrings(replacements)`
+### `replaceStrings(replacements)`
 
 Given a set of replacements (specified as a tuple of [string, string]`), this transform will replace the first item in the tupel with the second item.
 
 ```tsx
+import {transform, replaceStrings} from '@shopify/ast-transforms/javascript';
+
 const initial = `
   foo = 'foo';
   bar = 'bar';
@@ -108,11 +107,13 @@ const result = await transform(
 console.log(result);
 ```
 
-#### `replaceJsxBody(parent, children)`
+### `replaceJsxBody(parent, children)`
 
 Given the parent JSX string, will find and wrap a tree of JSX, beginning with the element that maches the second argument string. This is particularly useful for add React [providers](https://reactjs.org/docs/context.html#contextprovider).
 
 ```tsx
+import {transform, replaceJsxBody} from '@shopify/ast-transforms/javascript';
+
 const initial = `
   <Foo><Baz>{qux}</Baz></Foo>
 `;
@@ -122,11 +123,16 @@ const result = await transform(initial, replaceJsxBody(`<Bar></Bar>`, 'Baz'));
 console.log(result); // <Foo><Bar><Baz>{qux}</Baz></Bar></Foo>;
 ```
 
-#### `addVariableDeclaratorProps(prop)`
+### `addVariableDeclaratorProps(prop)`
 
 Adds a new local variable destructured from `this.props`;
 
 ```tsx
+import {
+  transform,
+  addVariableDeclaratorProps,
+} from '@shopify/ast-transforms/javascript';
+
 const initial = `const { prop1, prop2 } = this.props;`;
 
 const result = await transform(initial, addVariableDeclaratorProps('prop3'));
@@ -134,11 +140,13 @@ const result = await transform(initial, addVariableDeclaratorProps('prop3'));
 console.log(result); // const { prop1, prop2, prop3 } = this.props;
 ```
 
-#### `addInterface(interface)`
+### `addInterface(interface)`
 
 Adds a new Typescript interface from the first argument string. If the interface already exists, it will be merged with the new one.
 
 ```tsx
+import {transform, addInterface} from '@shopify/ast-transforms/javascript';
+
 const initial = `interface Foo {
   bar?: Baz
 }
@@ -158,3 +166,7 @@ const result = await transform(
 // };
 console.log(result);
 ```
+
+## Markdown
+
+### `addReleaseToChangelog(object)`
