+++ packages/ast-utilities/javascript.d.ts
@@ -1,0 +1,1 @@
+export * from './dist/javascript';
