+++ packages/ast-utilities/src/javascript/addComponentProps/addComponentProps.ts
