+++ packages/ast-utilities/src/javascript/addComponentProps/index.ts
