+++ packages/ast-utilities/src/javascript/addComponentProps/tests/addComponentProps.test.ts
