+++ packages/ast-utilities/src/javascript/addVariableDeclaratorProps/addVariableDeclaratorProps.ts
