+++ packages/ast-utilities/src/javascript/addVariableDeclaratorProps/index.ts
