+++ packages/ast-utilities/src/javascript/addVariableDeclaratorProps/tests/addVariableDeclaratorProps.test.ts
