+++ packages/ast-utilities/src/markdown/transforms/addReleaseToChangelog/index.ts
@@ -1,0 +1,3 @@
+import addReleaseToChangelog from './addReleaseToChangelog';
+
+export default addReleaseToChangelog;
