--- packages/ast-utilities/src/markdown/transforms/updateChangelog/index.ts
@@ -1,3 +1,0 @@
-import updateChangelog from './updateChangelog';
-
-export default updateChangelog;
