+++ packages/ast-utilities/src/matchers.ts
@@ -4,8 +4,6 @@ import {
   RECEIVED_COLOR as receivedColor,
 } from 'jest-matcher-utils';
 
-import {format} from './test/utilities';
-
 declare global {
   namespace jest {
     interface Matchers<R> {
@@ -40,3 +38,21 @@ expect.extend({
 });
 
 export {};
+
+function format(str: string) {
+  return (
+    str
+      .split('\n')
+      .map(line => line.trim())
+      .join('\n')
+      // outdent
+      .replace(/(^\s+|\s+$)/g, '')
+      // replace line breaks
+      .replace(/(\r\n|\n|\r)/gm, '')
+      // replace spaces in curly braces
+      .replace(/{ /g, '{')
+      .replace(/ }/g, '}')
+      // replace double with single quotes
+      .replace(/"/g, `'`)
+  );
+}
